fixed Window, now errors are gone

+++ cpp-atm/CardNumberInput.cpp
@@ -16,24 +16,24 @@ void CardNumberInput::preExecute()
 
 void CardNumberInput::removeSymbol()
 {
-	if (buffer.size() % 4 == 0)
-		ConsoleUtils::setCursorPosition(xpos + 1 + (buffer.size()) + buffer.size() / multipliter * 3 - 3, ypos + 1);
+	if ((int)buffer.size() % 4 == 0)
+		ConsoleUtils::setCursorPosition(xpos + 1 + ((int)buffer.size()) + (int)buffer.size() / multipliter * 3 - 3, ypos + 1);
 	else
-		ConsoleUtils::setCursorPosition(xpos + 1 + (buffer.size()) + buffer.size() / multipliter * 3, ypos + 1);
-	buffer.resize(buffer.size() - 1);
+		ConsoleUtils::setCursorPosition(xpos + 1 + ((int)buffer.size()) + (int)buffer.size() / multipliter * 3, ypos + 1);
+	buffer.resize((int)buffer.size() - 1);
 	std::cout << " ";
-	ConsoleUtils::setCursorPosition(xpos + 1 + (buffer.size()) + buffer.size() / multipliter * 3 + 1, ypos + 1);
+	ConsoleUtils::setCursorPosition(xpos + 1 + ((int)buffer.size()) + (int)buffer.size() / multipliter * 3 + 1, ypos + 1);
 }
 
 void CardNumberInput::addSymbol(SHORT key)
 {
 	int multipliter = 4;
-	if (key >= 48 && key <= 57 && buffer.size() < length)
+	if (key >= 48 && key <= 57 && (int)buffer.size() < length)
 	{
 		buffer += std::to_string((keyToIntValue(key)));
 		std::cout << keyToIntValue(key);
-		if (buffer.size() < 15)
-			ConsoleUtils::setCursorPosition(xpos + 1 + (buffer.size()) + 1 + buffer.size() / multipliter * 3, ypos + 1);
+		if ((int)buffer.size() < 15)
+			ConsoleUtils::setCursorPosition(xpos + 1 + ((int)buffer.size()) + 1 + (int)buffer.size() / multipliter * 3, ypos + 1);
 	}
 }
 
@@ -44,7 +44,7 @@ int CardNumberInput::keyToIntValue(SHORT key)
 
 int CardNumberInput::onReturn()
 {
-	if (buffer.size() == length)
+	if ((int)buffer.size() == length)
 		return 0;
 	return -1;
 }
