+++ cpp-atm/PinInput.cpp
@@ -17,10 +17,10 @@ void PinInput::preExecute()
 
 void PinInput::removeSymbol()
 {
-	ConsoleUtils::setCursorPosition(xpos + 2 + (buffer.size() - 1) * multipliter, ypos + 1);
-	buffer.resize(buffer.size() - 1);
+	ConsoleUtils::setCursorPosition(xpos + 2 + ((int)buffer.size() - 1) * multipliter, ypos + 1);
+	buffer.resize((int)buffer.size() - 1);
 	std::cout << " ";
-	ConsoleUtils::setCursorPosition(xpos + 2 + (buffer.size()) * multipliter, ypos + 1);
+	ConsoleUtils::setCursorPosition(xpos + 2 + ((int)buffer.size()) * multipliter, ypos + 1);
 }
 
 void PinInput::addSymbol(SHORT key)
@@ -28,11 +28,11 @@ void PinInput::addSymbol(SHORT key)
 	if (key >= 48 && key <= 57 && buffer.size() < 4)
 	{
 		//can i delete this???
-		ConsoleUtils::setCursorPosition(xpos + 2 + (buffer.size()) * multipliter, ypos + 1);
+		ConsoleUtils::setCursorPosition(xpos + 2 + ((int)buffer.size()) * multipliter, ypos + 1);
 		buffer += std::to_string((keyToIntValue(key)));
 		std::cout << keyToIntValue(key);
-		if (buffer.size() < 3)
-			ConsoleUtils::setCursorPosition(xpos + 2 + (buffer.size()) * multipliter, ypos + 1);
+		if ((int)buffer.size() < 3)
+			ConsoleUtils::setCursorPosition(xpos + 2 + ((int)buffer.size()) * multipliter, ypos + 1);
 	}
 }
 
@@ -43,7 +43,7 @@ int PinInput::keyToIntValue(SHORT key)
 
 int PinInput::onReturn()
 {
-	if (buffer.size() == length)
+	if ((int)buffer.size() == length)
 		return 0;
 	return -1;
 }
