+++ cpp-atm/Utility.h
@@ -12,7 +12,7 @@ public:
 	static string encryptPassword(const string& pass) {
         unsigned int hash = 0;
 
-        const unsigned int VALUE = pass.length();
+        const unsigned int VALUE = (unsigned int)pass.length();
         for (auto Letter : pass)
         {
             srand(VALUE * Letter);
